# Adaptive in-memory sort (~2x faster) (#5879)