@yjshen 2023-04-30
> Previously ExternalSorter would sort every batch it receives, but then perform an in-memory sort by concatenating all the inputs together.
>
> This PR modifies ExternalSorter to not presort batches

I think that sorting the batch while it is still in the cache, immediately after its generation, can effectively utilize the cache and enhance performance?