@tustvold 2023-04-30
> can effectively utilize the cache and enhance performance

At least empirically with the benchmarks I have available, the cost of sorting and therefore copying all the values far outweighs any cache locality effects.