@alamb 2023-04-30
I ran the `sort` benchmarks (shout out to @jaylmiller for adding them) on this branch https://github.com/apache/arrow-datafusion/tree/main/benchmarks#parquet-benchmarks

They certainly look very promising:

```
--------------------
Benchmark sort.json
--------------------
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━━━┳━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃        sort ┃       sort ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━━━╇━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ Qsort utf8   │  80037.05ms │ 55724.44ms │ +1.44x faster │
│ Qsort int    │  96166.00ms │ 69435.53ms │ +1.38x faster │
│ Qsort        │  85487.77ms │ 57109.82ms │ +1.50x faster │
│ decimal      │             │            │               │
│ Qsort        │ 103824.30ms │ 78792.01ms │ +1.32x faster │
│ integer      │             │            │               │
│ tuple        │             │            │               │
│ Qsort utf8   │  80706.73ms │ 63156.00ms │ +1.28x faster │
│ tuple        │             │            │               │
│ Qsort mixed  │  97100.16ms │ 68542.10ms │ +1.42x faster │
│ tuple        │             │            │               │
└──────────────┴─────────────┴────────────┴───────────────┘
```